moved styles.ts to appropriate folder
