+++ src/components/tile_card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import styles from './styles';
+import styles from '../util/styles';
 import { TileCardProps } from '../util/interface';
 
 const TileWrapper = styled.div`
